refactor: rename utils to core
