+++ src/cogs/admin.py
@@ -6,7 +6,7 @@ from typing import TYPE_CHECKING
 from discord import app_commands
 from discord.ext.commands import Cog  # pyright: ignore[reportMissingTypeStubs]
 
-from utils import config
+from core import config
 
 if TYPE_CHECKING:
     from discord import Interaction
