+++ src/core/temporary_cache.py
